+++ app/src/main/java/com/ing/software/ticketapp/OCR/OcrVars.java
@@ -8,6 +8,4 @@ public class OcrVars {
 
     static final boolean ISDEBUGENABLED = true;
     static final String amountString = "TOTALE";
-
-
 }
